+++ packages/reference/src/assets/WrappedAssetCard/FetchingWrappedAssetCard.tsx
@@ -30,6 +30,14 @@ type FetchingWrappedAssetCardProps = {
   renderDragHandle?: RenderDragFn;
   renderCustomCard?: CustomCardRenderer;
   renderCustomMissingEntityCard?: RenderCustomMissingEntityCard;
+  addReferenceToRelease?: (
+    reference: Asset,
+    localeCode?: string,
+    options?: {
+      openModalForVersionSelection?: boolean;
+      skipNestedReferencesPrompt?: boolean;
+    },
+  ) => Promise<void>;
 };
 
 export function FetchingWrappedAssetCard(props: FetchingWrappedAssetCardProps) {
@@ -49,6 +57,15 @@ export function FetchingWrappedAssetCard(props: FetchingWrappedAssetCardProps) {
     locales: props.sdk.locales,
     isReference: true,
   });
+
+  const onAddToRelease = () => {
+    if (asset && props.addReferenceToRelease) {
+      void props.addReferenceToRelease(asset, props.sdk.field.locale, {
+        openModalForVersionSelection: true,
+        skipNestedReferencesPrompt: true,
+      });
+    }
+  };
 
   React.useEffect(() => {
     if (asset) {
@@ -114,6 +131,7 @@ export function FetchingWrappedAssetCard(props: FetchingWrappedAssetCardProps) {
       releaseStatusMap,
       release: props.sdk.release as ReleaseV2Props | undefined,
       releaseEntityStatus,
+      onAddToRelease,
     };
 
     if (status === 'loading') {