+++ packages/rating/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.4](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rating@1.7.3...@contentful/field-editor-rating@1.7.4) (2026-01-22)
+
+**Note:** Version bump only for package @contentful/field-editor-rating
+
+## [1.7.3](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rating@1.7.2...@contentful/field-editor-rating@1.7.3) (2026-01-19)
+
+**Note:** Version bump only for package @contentful/field-editor-rating
+
 ## [1.7.2](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rating@1.7.1...@contentful/field-editor-rating@1.7.2) (2025-11-17)
 
 ### Bug Fixes